add suggested changes
