+++ middleware/settings.py
@@ -156,36 +156,6 @@ ASGI_APPLICATION = "middleware.asgi.application"
 CHANNEL_LAYERS = {"default": {"BACKEND": "channels.layers.InMemoryChannelLayer"}}
 
 
-CARE_URL = env("CARE_URL")
-
-JWKS = JsonWebKey.import_key_set(
-    json.loads(base64.b64decode(env("JWKS_BASE64", default=generate_encoded_jwks())))
-)
-
-
-REDIS_OBSERVATIONS_KEY = env("REDIS_OBSERVATIONS_KEY")
-
-
-FACILITYID = "81092ced-8720-44cb-b4c5-3f0ad0540153"
-HOST_NAME = "appliances-trek-termination-hotel.trycloudflare.com"
-CSRF_TRUSTED_ORIGINS = ["https://appliances-trek-termination-hotel.trycloudflare.com"]
-UPDATE_INTERVAL = 60 * 60 * 1000
-DEFAULT_LISTING_LIMIT = 10
-CELERY_BROKER_URL = (
-    "redis://127.0.0.1:6379"  # CELERY_ACCEPT_CONTENT = ["application/json"]
-)
-# CELERY_RESULT_SERIALIZER = "json"
-# CELERY_TASK_SERIALIZER = "json"
-# CELERY_RESULT_BACKEND = "django-db"
-# CELERY_TIMEZONE = "Asia/Karachi"
-
-
-CELERY_BEAT_SCHEDULER = "django_celery_beat.schedulers:DatabaseScheduler"
-
-
-# CELERY_BROKER_URL = config("CELERY_BROKER_REDIS_URL", default="redis://localhost:6379")
-
-
 CACHES = {
     "default": {
         "BACKEND": "django_redis.cache.RedisCache",
@@ -195,3 +165,23 @@ CACHES = {
         },
     }
 }
+
+
+# Configs
+CARE_URL = env("CARE_URL")
+FACILITYID = env("FACILITYID")
+CARE_JWK_URL = env("CARE_JWK_URL")
+JWKS = JsonWebKey.import_key_set(
+    json.loads(base64.b64decode(env("JWKS_BASE64", default=generate_encoded_jwks())))
+)
+CELERY_BROKER_URL = "redis://127.0.0.1:6379"
+CELERY_BEAT_SCHEDULER = "django_celery_beat.schedulers:DatabaseScheduler"
+
+HOST_NAME = env("HOST_NAME")
+CSRF_TRUSTED_ORIGINS = ["https://appliances-trek-termination-hotel.trycloudflare.com"]
+
+
+# Observations
+REDIS_OBSERVATIONS_KEY = env("REDIS_OBSERVATIONS_KEY")
+UPDATE_INTERVAL = env("UPDATE_INTERVAL")
+DEFAULT_LISTING_LIMIT = env("DEFAULT_LISTING_LIMIT")
