+++ middleware/tasks.py
@@ -6,22 +6,22 @@ from django.conf import settings
 import logging
 
 from middleware.models import Asset, AssetClasses
-from middleware.serializers.observation import (
+from middleware.types.observations import (
     DailyRoundObservation,
 )
+
 from middleware.utils.observation_utils import get_vitals_from_observations
-from middleware.utils.utils import _get_headers, get_patient_id
+from middleware.utils.utils import (
+    _get_headers,
+    file_automated_daily_rounds,
+    get_patient_id,
+)
 
 logger = logging.getLogger(__name__)
 
 @shared_task
-def test_celery_task():
-    return print("celery is working ")
-
-
-@shared_task
 def retrieve_asset_config():
-    print("retreving assets")
+    logger.info("Started Retrieving Assets Task")
     response = requests.get(
         f"{settings.CARE_URL}asset_config/?middleware_hostname={settings.HOST_NAME}",
         headers=_get_headers(),
@@ -29,20 +29,20 @@ def retrieve_asset_config():
 
     response.raise_for_status()
     data = response.json()
-    print("data is", data)
     existing_asset_ids = list(
         Asset.objects.filter(deleted=False).values_list("id", flat=True)
     )
-    print("existing", existing_asset_ids)
+    logger.info("Existing  Asset ids: %s", existing_asset_ids)
     missing_asset_ids = [
         asset["id"] for asset in data if UUID(asset["id"]) not in existing_asset_ids
     ]
-    print("missing_asset_ids", missing_asset_ids)
+
+    logger.info("Missing  Asset ids: %s", missing_asset_ids)
 
     # Mark missing assets as deleted
     deleted_count = Asset.objects.filter(id__in=missing_asset_ids).update(deleted=True)
 
-    print("Deleted assets count:", deleted_count)
+    logger.info("Deleted assets count: %s ", deleted_count)
 
     for asset in data:
         # Implement logic to create or update assets based on your model
@@ -53,13 +53,11 @@ def retrieve_asset_config():
 
 @shared_task
 def automated_daily_rounds():
-    print("Automated daily rounds")
-
+    logger.info("Started Automated daily rounds")
     monitors = Asset.objects.filter(type=AssetClasses.HL7MONITOR.name, deleted=False)
-    print("Found %s monitors", len(monitors))
-
+    logger.info("Found %s monitors", len(monitors))
     for monitor in monitors:
-        print("Processing monitor %s", monitor.id)
+        logger.info("Processing Monitor having id: %s", monitor.id)
         consultation_id, patient_id, bed_id, asset_beds = get_patient_id(
             external_id=monitor.id
         )
@@ -70,4 +68,8 @@ def automated_daily_rounds():
         vitals: Optional[DailyRoundObservation] = get_vitals_from_observations(
             ip_address=monitor.ip_address
         )
-        print("vitals", vitals)
+        logger.info("Vitals for Monitor having id:%s  is: %s", monitor.id, vitals)
+
+        file_automated_daily_rounds(
+            consultation_id=consultation_id, asset_id=monitor.id, vitals=vitals
+        )
