+++ middleware/utils/utils.py
@@ -1,3 +1,4 @@
+import logging
 from typing import Dict, List
 from uuid import UUID
 import requests
@@ -7,7 +8,11 @@ from middleware.token_generator import generate_jwt
 from typing import List, Dict, TypeVar, Any
 from pydantic import BaseModel
 
+from middleware.types.observations import DailyRoundObservation
+
 T = TypeVar("T", bound=BaseModel)
+logger = logging.getLogger(__name__)
+
 
 def _get_headers(claims: dict = None) -> dict:
     return {
@@ -47,10 +52,27 @@ def get_patient_id(external_id: UUID):
     )
     response.raise_for_status()
     data = response.json()
-    print("data is ", data)
     return (
         data.get("consultation_id"),
         data.get("patient_id"),
         data.get("bed_id"),
         data.get("asset_beds"),
     )
+
+
+def file_automated_daily_rounds(
+    consultation_id: UUID, asset_id: UUID, vitals: DailyRoundObservation
+):
+    response = requests.post(
+        f"{settings.CARE_URL}consultation/{consultation_id}/daily_rounds/",
+        data=vitals,
+        headers=_get_headers(claims={"asset_id": str(asset_id)}),
+    )
+    if response.status_code != 201:
+        logger.error(
+            "Failed to file the daily round for the consultation: $%s and asset:%s",
+            consultation_id,
+            asset_id,
+        )
+        return
+    response.raise_for_status()
